+++ NEWS.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* New: [View::hide] temporarily hides the view, but allows it to pop back when
-  the model is next updated.
+* New: [View::clear] temporarily removes the progress view if it is drawn, but
+  allows it to pop back when the model is next updated.
 
 * New: [Options] can be constructed as a `static` value: there's a new
   `const fn` [Options::new] constructor and the functions to set fields are also